requirements.txt
astroid==2.0.4
autopep8==1.3.5
Django==2.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-tinymce==2.7.0
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
pycodestyle==2.4.0
pylint==2.1.1
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
